Extract UberBottomSheetListItem
